// LLD/Machine-coding/Ride-Sharing/helper/enums.ts
export enum Gender {
    MALE = "MALE",
    FEMALE = "FEMALE",
    OTHER = "OTHER"
}

export enum RideAssetType {
    TWO_WHEELER = "TWO_WHEELER",
    FOUR_WHEELER = "FOUR_WHEELER"
}

export enum DriverStatus {
    ONLINE = "ONLINE",
    OFFLINE = "OFFLINE",
    ON_TRIP = "ON_TRIP"
}

export enum TripStatusType {
    ON_GOING = "ON_GOING",
    COMPLETE = "COMPLETE"
}